Move cursor controller to service layer
