+++ server/service/src/lib.rs
@@ -14,6 +14,7 @@ pub mod auth_data;
 pub mod barcode;
 pub mod clinician;
 mod common_stock;
+pub mod cursor_controller;
 pub mod dashboard;
 pub mod display_settings_service;
 pub mod document;
@@ -275,8 +276,11 @@ fn check_location_exists(
     store_id: &str,
     location_input: &Option<NullableUpdate<String>>,
 ) -> Result<bool, RepositoryError> {
-    let Some(NullableUpdate{ value: Some(location_id) }) = location_input else {
-        return Ok(true)
+    let Some(NullableUpdate {
+        value: Some(location_id),
+    }) = location_input
+    else {
+        return Ok(true);
     };
     let count = LocationRepository::new(connection).count(Some(
         LocationFilter::new()
